# flyio-admin/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "flyio-admin API",
    "description": "Application and Orchestration layer API for Flyio.ai — CMS, Blog, Trips, Knowledge Base, Web Scraper coordination, and LLM Workflow Orchestration.",
    "version": "1.0.0",
    "contact": {
      "name": "Flyio.ai Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT token acquired via POST /api/admin/login"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/knowledge-base": {
      "get": {
        "summary": "List indexed knowledge base sources",
        "tags": ["Knowledge Base & Scraper"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Paginated knowledge base sources" }
        }
      },
      "delete": {
        "summary": "Delete entries from knowledge base by source_url or chunk id",
        "tags": ["Knowledge Base & Scraper"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "source_url", "in": "query", "schema": { "type": "string" } },
          { "name": "id", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Deleted from knowledge base" }
        }
      }
    },
    "/api/admin/knowledge-base/chunks": {
      "get": {
        "summary": "List all chunks for a source URL",
        "tags": ["Knowledge Base & Scraper"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "source_url", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Chunks list" }
        }
      }
    },
    "/api/admin/crawl-history": {
      "get": {
        "summary": "Permanent audit log of all crawl attempts",
        "tags": ["Knowledge Base & Scraper"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Paginated crawl history" }
        }
      }
    },
...
  }
}